requirements.txt
asgiref==3.8.1
Django==4.2.7
Pillow==10.0.1
psycopg2-binary==2.9.10
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.2